refactor(#537): split MapLayersListModal into focused files per Zach's review
Reduces MapLayersListModal.tsx from 1206 lines to 244 by extracting:
- MapLayersTreeItem.tsx — reusable tree node component
- useMapLayersModal.ts  — all hook/state/effect logic
- MarkersLayerSection.tsx, StationsLayerSection.tsx,
  PolygonsLayerSection.tsx, TileLayersSection.tsx,
  KmlLayersSection.tsx  — one file per layer type

All files are under Zach's 400-500 line guideline. No logic changed.
